move tests for removed Item methods into ItemService
